+++ backend/views/article/index.php
@@ -41,7 +41,14 @@ $this->params['breadcrumbs'][] = $this->title;
                     return $model->author->username;
                 }
             ],
-            'status',
+            [
+                'class'=>\common\components\grid\EnumColumn::className(),
+                'attribute'=>'status',
+                'enum'=>[
+                    \Yii::t('backend', 'Disabled'),
+                    \Yii::t('backend', 'Enabled')
+                ],
+            ],
             'published_at:datetime',
             'created_at:datetime',