+++ backend/views/widget-carousel/index.php
@@ -4,7 +4,7 @@ use yii\helpers\Html;
 use yii\grid\GridView;
 
 /* @var $this yii\web\View */
-/* @var $searchModel backend\\models\search\WidgetCarouselSearch */
+/* @var $searchModel backend\models\search\WidgetCarouselSearch */
 /* @var $dataProvider yii\data\ActiveDataProvider */
 
 $this->title = Yii::t('backend', 'Widget Carousels');
@@ -26,7 +26,14 @@ $this->params['breadcrumbs'][] = $this->title;
 
             'id',
             'alias',
-            'status',
+            [
+                'class'=>\common\components\grid\EnumColumn::className(),
+                'attribute'=>'status',
+                'enum'=>[
+                    \Yii::t('backend', 'Disabled'),
+                    \Yii::t('backend', 'Enabled')
+                ],
+            ],
 
             [
                 'class' => 'yii\grid\ActionColumn',